pyproject.toml
[tool.poetry]
name = "final project"
version = "0.1.0"
description = "final project from the next level course from ArjanCodes"
authors = ["Joao Martins"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.2"
dash = "^2.10.2"
dash-bootstrap-components = "^1.4.1"
black = "^23.3.0"
pre-commit = "^3.3.2"
flake8 = "^6.0.0"
babel = "^2.12.1"
plotly = "^5.15.0"
pandas-stubs = "^2.0.2.230605"
python-i18n = "^0.3.9"
ipython = "^8.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
